Ниты: hms без 32-битного overflow + запрет amount_range.min == 0
Две мелкие правки (пропущены codex-ревью):

  * hms(): total теперь long long, а не long. На Windows (LLP64) long —
    32 бита, и модельный аптайм (реальные секунды * time_scale) при
    ускоренном времени переполнял бы его за часы прогона, печатая
    отрицательное HH:MM:SS в шапке дашборда.

  * validate(): amount_range требует min >= 1 (было >= 0). Суммы из
    диапазона идут в снятие/внесение, а те при amount <= 0 возвращают
    InvalidAmount — конфиг с min == 0 порождал клиентов с гарантированно
    проваленной операцией. CheckBalance ставит 0 сам, в обход диапазона.

Тест: config_rejects_zero_amount_min. Все дефолты (min=50000/20000)
остаются валидными. Сборка g++14 + 64 теста зелёные.

+++ src/console/LiveRenderer.cpp
@@ -15,7 +15,10 @@ namespace {
 
 // Форматирует секунды как HH:MM:SS.
 std::string hms(double seconds) {
-    long total = static_cast<long>(seconds < 0 ? 0 : seconds);
+    // long long (>= 64 бит везде), а не long: на Windows (LLP64) long — 32 бита,
+    // и модельный аптайм (реальные секунды * time_scale) при заметном ускорении
+    // времени переполнил бы его за часы прогона, дав отрицательное HH:MM:SS.
+    long long total = static_cast<long long>(seconds < 0 ? 0 : seconds);
     std::ostringstream os;
     os << std::setfill('0') << std::setw(2) << (total / 3600) << ':'
        << std::setw(2) << ((total % 3600) / 60) << ':' << std::setw(2) << (total % 60);
